[torch.library] Add ability to create library fragments
In C++ we have TORCH_LIBRARY_FRAGMENT. This PR adds the same
functionality to the Python torch.library API.

The motivation for this is: for the simple custom op API, we don't want
users to need to deal with Library objects. One way to hide this from
users is to create library fragments.

Test Plan:
- tests that you can create multiple fragments and def+impl operators on each.

ghstack-source-id: 513debc9d23a1d9cf9f9ac88a541f8a9d15d2b65
Pull Request resolved: https://github.com/pytorch/pytorch/pull/98439